+++ google/resource_compute_instance_template.go
@@ -52,8 +52,6 @@ func resourceComputeInstanceTemplate() *schema.Resource {
 					}
 					return
 				},
-				Deprecated: "Use the random provider instead. See migration instructions at " +
-					"https://github.com/terraform-providers/terraform-provider-google/issues/1054#issuecomment-377390209",
 			},
 
 			"disk": &schema.Schema{